Extract class SimulationTime out of the Simulation class
